brainTumor_YOLO/Brain-Tumor-Detection/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
flask-cors==3.0.10
Pillow==9.0.0
ultralytics==8.0.0
torch==2.0.0
opencv-python==4.7.0.72
numpy==1.24.0
<<<<<<< HEAD
matplotlib==3.7.0 
gunicorn
=======
matplotlib==3.7.0
gunicorn==20.1.0
python-dateutil==2.8.2 
>>>>>>> ee0f3d1 (feat: Add deployment configuration files and update requirements)
